+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 - PowerShell exploiter plugin. #4256
 - Log4Shell exploiter plugin. #4257
 - SNMP exploiter plugin. #4258
+- Zerologon exploiter plugin. #4259
 
 ### Security